cleaning up process by creating handle.go
